Call standalone script as console script.
setuptools allows creating executable scripts from your python files.
You won't have to type `python path/to/package.py` any more. We define
where the script exists and how to call it in setup.py, so now on the
command line you'd go:
$ evt_standalone.

This requires to put what was in the main body of the file into some
function, usually we call it main.